:recycle: Rename to monthly_calendar_layout
